Edit: Moved repositories on chooseGame to domain and infra layers
