+++ src/modules/chooseGame/infra/repositories/factories/makeGamesRepository.ts
@@ -1,5 +1,5 @@
+import IGamesRepository from 'modules/chooseGame/domain/repositories/IGamesRepository';
 import GamesRepository from 'modules/chooseGame/infra/repositories/GamesRepository';
-import IGamesRepository from 'modules/chooseGame/repositories/IGamesRepository';
 
 export default function makeGamesRepository(): IGamesRepository {
   return new GamesRepository();
