+++ src/modules/chooseGame/infra/repositories/factories/makePlayersRepository.ts
@@ -1,5 +1,5 @@
+import IPlayersRepository from 'modules/chooseGame/domain/repositories/IPlayersRepository';
 import PlayersRepository from 'modules/chooseGame/infra/repositories/PlayersRepository';
-import IPlayersRepository from 'modules/chooseGame/repositories/IPlayersRepository';
 
 export default function makePlayersRepository(): IPlayersRepository {
   return new PlayersRepository();
