+++ src/modules/chooseGame/infra/repositories/factories/makeUsersRepository.ts
@@ -1,5 +1,5 @@
+import IUsersRepository from 'modules/chooseGame/domain/repositories/IUsersRepository';
 import UsersRepository from 'modules/chooseGame/infra/repositories/UsersRepository';
-import IUsersRepository from 'modules/chooseGame/repositories/IUsersRepository';
 
 export default function makeUsersRepository(): IUsersRepository {
   return new UsersRepository();
