+++ src/modules/chooseGame/services/GetUserService.ts
@@ -1,4 +1,4 @@
-import IUsersRepository from 'modules/chooseGame/repositories/IUsersRepository';
+import IUsersRepository from 'modules/chooseGame/domain/repositories/IUsersRepository';
 import IUser from 'shared/domain/entities/IUser';
 
 interface IExecute {
