+++ src/modules/chooseGame/services/factories/makeFetchPlayersService.ts
@@ -1,4 +1,4 @@
-import makePlayersRepository from 'modules/chooseGame/repositories/factories/makePlayersRepository';
+import makePlayersRepository from 'modules/chooseGame/infra/repositories/factories/makePlayersRepository';
 import FetchPlayersService from 'modules/chooseGame/services/FetchPlayersService';
 
 export default function makeFetchPlayersService(): FetchPlayersService {
