+++ src/modules/chooseGame/services/factories/makeGetUserService.ts
@@ -1,4 +1,4 @@
-import makeUsersRepository from 'modules/chooseGame/repositories/factories/makeUsersRepository';
+import makeUsersRepository from 'modules/chooseGame/infra/repositories/factories/makeUsersRepository';
 import GetUserService from 'modules/chooseGame/services/GetUserService';
 
 export default function makeGetUserService(): GetUserService {
