+++ src/modules/chooseGame/services/factories/makeUpdateUserService.ts
@@ -1,4 +1,4 @@
-import makeUsersRepository from 'modules/chooseGame/repositories/factories/makeUsersRepository';
+import makeUsersRepository from 'modules/chooseGame/infra/repositories/factories/makeUsersRepository';
 import UpdateUserService from 'modules/chooseGame/services/UpdateUserService';
 
 export default function makeUpdateUserService(): UpdateUserService {
